requirements.txt
setuptools>=68.0.0
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.1
django-cors-headers==4.4.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
cryptography==42.0.8
djangorestframework-simplejwt>=5.3.0

PyJWT==2.8.0
requests==2.32.3
gunicorn==21.2.0
django-filter==23.5
django-extensions==3.2.3
twilio==9.3.0  # For SMS
firebase-admin==6.5.0  # For Firebase Push Notifications
kafka-python>=2.0.2  # For Kafka producer/consumer
django-environ
drf-yasg==1.21.7
channels==4.1.0  # For WebSockets
channels_redis==4.2.0  # Redis backend
daphne==4.2.1  # ASGI server for WebSockets

# Testing dependencies
pytest==8.0.0
pytest-django==4.8.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
pytest-mock==3.12.0